requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
psycopg2-binary==2.9.9
python-dotenv==1.0.0
Pillow==10.1.0
aiofiles==23.2.1
